leetcode-ai-planner/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
google-generativeai==0.3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
redis==5.0.1
langgraph==0.0.25
langchain==0.1.0
openpyxl==3.1.2
pydantic==2.5.0
PyPDF2==3.0.1